refactor(DrawerTitle): move compoent to its own file
