+++ src/components/drawers/Settings/Projects/ProjectDrawer.tsx
@@ -4,7 +4,7 @@ import { FormControlLabel, Typography } from '@mui/material'
 import { RhfSwitch } from 'mui-rhf-integration'
 import { zodResolver } from '@hookform/resolvers/zod'
 
-import DrawerTitle from '~components/titles/DrawerTitle'
+import DrawerTitle from '~/components/titles/DrawerTitle/DrawerTitle'
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
 import { projectSchema, ProjectType } from '~types/project'
 import BcaTextField from '~/components/input/BcaTextField/BcaTextField'
