+++ src/components/drawers/Settings/RubroMaterial/RubroMaterialsDrawer.tsx
@@ -5,7 +5,7 @@ import { Typography } from '@mui/material'
 
 import { rubroMaterialSchema, RubroMaterialType } from '~types/rubro-material'
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
-import DrawerTitle from '~components/titles/DrawerTitle'
+import DrawerTitle from '~/components/titles/DrawerTitle/DrawerTitle'
 import ButtonGroup from '~components/buttons/button-group'
 import BcaTextField from '~/components/input/BcaTextField/BcaTextField'
 import { useGetAllMaterialsQuery } from '~redux/api/bca-backend/parametros/materialsSlice'
