+++ src/components/drawers/Settings/Suppliers/SupplierDrawer.tsx
@@ -4,7 +4,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { Typography } from '@mui/material'
 
 import { supplierSchema, SupplierType } from '~types/supplier'
-import DrawerTitle from '~components/titles/DrawerTitle'
+import DrawerTitle from '~/components/titles/DrawerTitle/DrawerTitle'
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
 import BcaTextField from '~/components/input/BcaTextField/BcaTextField'
 import ButtonGroup from '~components/buttons/button-group'
