+++ src/components/drawers/Users/UsersDrawer.tsx
@@ -10,7 +10,7 @@ import {
     UserResponse,
     userResponseSchema,
 } from '~types/user'
-import DrawerTitle from '~components/titles/DrawerTitle'
+import DrawerTitle from '~/components/titles/DrawerTitle/DrawerTitle'
 import ButtonGroup from '~components/buttons/button-group'
 import {
     useCreateUserMutation,
