+++ src/pages/users/password/index.tsx
@@ -4,7 +4,7 @@ import { Box } from '@mui/material'
 import { useForm } from 'react-hook-form'
 
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
-import DrawerTitle from '~components/titles/DrawerTitle'
+import DrawerTitle from '~/components/titles/DrawerTitle/DrawerTitle'
 import ButtonGroup from '~components/buttons/button-group'
 import { passwordSchema, PasswordType } from '~types/user'
 import BcaTextField from '~/components/input/BcaTextField/BcaTextField'
